# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "EstadosFinancieros": {
        "description": "Estados estimados, auditables línea por línea.",
        "properties": {
          "anios": {
            "items": {
              "type": "integer"
            },
            "title": "Anios",
            "type": "array"
          },
          "id_caso": {
            "title": "Id Caso",
            "type": "string"
          },
          "lineas": {
            "additionalProperties": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "description": "Cada línea del estado, con un valor por año del horizonte",
            "title": "Lineas",
            "type": "object"
          }
        },
        "required": [
          "id_caso",
          "anios",
          "lineas"
        ],
        "title": "EstadosFinancieros",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Indicadores": {
        "properties": {
          "capital_intensity": {
            "title": "Capital Intensity",
            "type": "number"
          },
          "npv_musd": {
            "description": "Valor actual neto, millones de dólares",
            "title": "Npv Musd",
            "type": "number"
          },
          "payback_anios": {
            "title": "Payback Anios",
            "type": "number"
          },
          "tir": {
            "description": "Tasa interna de retorno, en tanto por uno",
            "title": "Tir",
            "type": "number"
          }
        },
        "required": [
          "npv_musd",
          "tir",
          "payback_anios",
          "capital_intensity"
        ],
        "title": "Indicadores",
        "type": "object"
      },
...
      "Problema": {
        "description": "Cuerpo de error. Explica qué pasó y cómo resolverlo.",
        "properties": {
          "detalle": {
            "title": "Detalle",
            "type": "string"
          },
          "restriccion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Restricción del cuadro de control que bloquea, si aplica",
            "title": "Restriccion"
          }
        },
        "required": [
          "detalle"
        ],
        "title": "Problema",
        "type": "object"
      },
...
      "ResultadoExportacion": {
        "properties": {
          "expira": {
            "format": "date-time",
            "title": "Expira",
            "type": "string"
          },
          "formato": {
            "enum": [
              "xlsx",
              "pdf"
            ],
            "title": "Formato",
            "type": "string"
          },
          "url_descarga": {
            "description": "URL temporal de lectura",
            "title": "Url Descarga",
            "type": "string"
          },
          "url_sharepoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Presente si el sitio de SharePoint está habilitado (R-44)",
            "title": "Url Sharepoint"
          }
        },
        "required": [
          "formato",
          "url_descarga",
          "expira"
        ],
        "title": "ResultadoExportacion",
        "type": "object"
      },
...
      "Tablero": {
        "properties": {
          "capex_por_etapa": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Capex Por Etapa",
            "type": "array"
          },
          "cascada": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Cascada",
            "type": "array"
          },
          "curva_costo_unitario": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Curva Costo Unitario",
            "type": "array"
          },
          "curva_produccion": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Curva Produccion",
            "type": "array"
          },
          "id_caso": {
            "title": "Id Caso",
            "type": "string"
          },
          "indicadores": {
            "$ref": "#/components/schemas/Indicadores"
          }
        },
        "required": [
          "id_caso",
          "indicadores"
        ],
        "title": "Tablero",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "name": "INVA · Servicio INVA-01-2026-182"
    },
    "description": "\nServicios de la Plataforma de Evaluación Económica · Modelo MINSUR.\n\nReproduce la lógica del modelo económico corporativo para que el área de\nProyectos evalúe escenarios de forma autónoma, con trazabilidad completa.\n\n**Autenticación.** Token de Microsoft Entra ID del tenant corporativo. El\nperfil se deriva de la reclamación `groups`; la plataforma no consulta la\nmembresía a Microsoft Graph.\n\n**Archivos.** Las plantillas se cargan y los reportes se descargan\ndirectamente contra el almacenamiento, mediante firmas de acceso compartido\nde corta vigencia. No atraviesan esta API.\n\n**Respuestas 501.** Indican una operación cuya implementación depende de un\ninsumo del cliente aún no recibido. El cuerpo incluye la restricción del\ncuadro de control que la bloquea.\n",
    "title": "Plataforma de Evaluación Económica · MINSUR",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/casos/{id_caso}/exportar": {
      "get": {
        "description": "Genera el reporte y devuelve una URL temporal de lectura.\n\nEl archivo se descarga directamente del almacenamiento, por el mismo\nmecanismo de firma que usa la carga: no atraviesa la puerta de enlace.\n\nCon `destino=sharepoint` publica ademas en la biblioteca documental del\nproyecto. Si el sitio aun no esta habilitado (R-44), la exportacion se\ncompleta y `url_sharepoint` queda vacia: es una limitacion de\nhabilitacion, no un defecto de la plataforma.\n\nEs la cuarta de las cinco comprobaciones obligatorias del pase.",
        "operationId": "exportar_api_casos__id_caso__exportar_get",
        "parameters": [
          {
            "in": "path",
            "name": "id_caso",
            "required": true,
            "schema": {
              "title": "Id Caso",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "destino",
            "required": false,
            "schema": {
              "default": "descarga",
              "pattern": "^(descarga|sharepoint)$",
              "title": "Destino",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "formato",
            "required": false,
            "schema": {
              "default": "xlsx",
              "pattern": "^(xlsx|pdf)$",
              "title": "Formato",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoExportacion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Exportar el tablero y los estados a Excel o PDF",
        "tags": [
          "Tablero"
        ]
      }
    },
...
    "/api/estados-financieros": {
      "get": {
        "description": "Cada linea con un valor por ano del horizonte detectado.\n\nLa auditabilidad linea por linea es lo que permite localizar una\ndiscrepancia contra el modelo de referencia en el bloque que la origina,\ny no solo constatar que el indicador final difiere.",
        "operationId": "estados_financieros_api_estados_financieros_get",
        "parameters": [
          {
            "in": "query",
            "name": "caso",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Caso"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadosFinancieros"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Estados financieros estimados, auditables linea por linea",
        "tags": [
          "Tablero"
        ]
      }
    },
...
    "/api/tablero": {
      "get": {
        "description": "Apertura comprometida por debajo de cinco segundos.\n\nEs la tercera de las cinco comprobaciones obligatorias de la ventana de\npase, que mide esa latencia contra el objetivo del alcance.",
        "operationId": "tablero_api_tablero_get",
        "parameters": [
          {
            "description": "Caso a mostrar",
            "in": "query",
            "name": "caso",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caso a mostrar",
              "title": "Caso"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tablero"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Indicadores, cascada y curvas del caso",
        "tags": [
          "Tablero"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "A través de API Management",
      "url": "/"
    }
  ]
}